+++ packages/react-components/react-combobox/src/components/Listbox/Listbox.tsx
@@ -5,6 +5,7 @@ import { useListboxStyles_unstable } from './useListboxStyles';
 import type { ListboxProps } from './Listbox.types';
 import { useListboxContextValues } from '../../contexts/useListboxContextValues';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
+import { useCustomStyleHooks_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * Listbox component: a standalone selection control, or the popup in a Combobox
@@ -14,6 +15,10 @@ export const Listbox: ForwardRefComponent<ListboxProps> = React.forwardRef((prop
   const contextValues = useListboxContextValues(state);
 
   useListboxStyles_unstable(state);
+
+  const { useListboxStyles_unstable: useCustomStyles } = useCustomStyleHooks_unstable();
+  useCustomStyles(state);
+
   return renderListbox_unstable(state, contextValues);
 });